requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-dotenv==1.0.1
python-multipart==0.0.19
easyocr==1.7.2
strands-agents==0.1.5
strands-agents-tools==0.1.2
openai==1.59.5
pydantic-settings==2.7.0
httpx==0.28.1
ddgs==9.13.0
